payments corrections

+++ db/migrate/20170913012134_change_data_type_for_price.rb
@@ -1,5 +1,5 @@
 class ChangeDataTypeForPrice < ActiveRecord::Migration
   def change 
-    change_column :products, :price, :decimal
+    change_column :products, :price, :integer
   end
 end
